requirements.txt
apache-airflow==2.10.4
apache-airflow-providers-postgres
apache-airflow-providers-openlineage
dbt-postgres
faker
pandas
sqlalchemy
psycopg2-binary
pytest
